fix(vacancy): address PR review comments and test fixes [BJS2-77435]
