+++ src/main/java/faang/school/projectservice/dto/vacancy/UpdateVacancyDto.java
@@ -1,34 +1,20 @@
-package faang.school.projectservice.dto;
+package faang.school.projectservice.dto.vacancy;
 
-import faang.school.projectservice.model.TeamRole;
 import faang.school.projectservice.model.WorkSchedule;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
 @NoArgsConstructor
-public class CreateVacancyDto {
-    @NotBlank
-    private String name;
+public class UpdateVacancyDto {
 
-    @NotBlank
     @Size(max = 3000)
     private String description;
-
-    @NotNull
-    private TeamRole position;
-
-    @NotNull
-    private Long projectId;
 
     @Positive
     private Double salary;
@@ -38,8 +24,4 @@ public class CreateVacancyDto {
 
     @Min(1)
     private Integer count;
-
-    private List<Long> requiredSkillIds;
-
-    private String coverImageKey;
 }
