+++ src/main/java/faang/school/projectservice/dto/vacancy/VacancyFilterDto.java
@@ -1,4 +1,4 @@
-package faang.school.projectservice.dto;
+package faang.school.projectservice.dto.vacancy;
 
 import faang.school.projectservice.model.TeamRole;
 import lombok.Data;
