+++ src/main/java/faang/school/projectservice/event/vacancy/DomainEventPublisher.java
@@ -1,4 +1,4 @@
-package faang.school.projectservice.event;
+package faang.school.projectservice.event.vacancy;
 
 public interface DomainEventPublisher {
     void publishEvent(DomainEvent event);
