+++ src/main/java/faang/school/projectservice/event/vacancy/LogOnlyPublisher.java
@@ -1,4 +1,4 @@
-package faang.school.projectservice.event;
+package faang.school.projectservice.event.vacancy;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Primary;
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Component;
 @Component
 @Slf4j
 @Primary
-//@Profile({"dev", "test"})
 public class LogOnlyPublisher implements DomainEventPublisher {
     @Override
     public void publishEvent(DomainEvent event) {
