+++ src/main/java/faang/school/projectservice/mapper/VacancyMapper.java
@@ -1,9 +1,9 @@
 package faang.school.projectservice.mapper;
 
-import faang.school.projectservice.dto.CreateVacancyDto;
-import faang.school.projectservice.dto.DetailedVacancyDto;
-import faang.school.projectservice.dto.UpdateVacancyDto;
-import faang.school.projectservice.dto.VacancyDto;
+import faang.school.projectservice.dto.vacancy.CreateVacancyDto;
+import faang.school.projectservice.dto.vacancy.DetailedVacancyDto;
+import faang.school.projectservice.dto.vacancy.UpdateVacancyDto;
+import faang.school.projectservice.dto.vacancy.VacancyDto;
 import faang.school.projectservice.model.Vacancy;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
