+++ src/main/java/faang/school/projectservice/service/VacancyService.java
@@ -1,12 +1,12 @@
 package faang.school.projectservice.service;
 
-import faang.school.projectservice.dto.CandidateDto;
-import faang.school.projectservice.dto.CreateCandidateDto;
-import faang.school.projectservice.dto.CreateVacancyDto;
-import faang.school.projectservice.dto.DetailedVacancyDto;
-import faang.school.projectservice.dto.UpdateVacancyDto;
-import faang.school.projectservice.dto.VacancyDto;
-import faang.school.projectservice.dto.VacancyFilterDto;
+import faang.school.projectservice.dto.candidate.CandidateDto;
+import faang.school.projectservice.dto.candidate.CreateCandidateDto;
+import faang.school.projectservice.dto.vacancy.CreateVacancyDto;
+import faang.school.projectservice.dto.vacancy.DetailedVacancyDto;
+import faang.school.projectservice.dto.vacancy.UpdateVacancyDto;
+import faang.school.projectservice.dto.vacancy.VacancyDto;
+import faang.school.projectservice.dto.vacancy.VacancyFilterDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
